Updated set_hvac_mode to track fan runtime from FAN_ONLY to FAN_ONLY
Data points are added every 30 seconds to prevent overfilling buffer and logs.

+++ app/include/thermostat.hpp
@@ -88,7 +88,7 @@ typedef struct
     bool thermostatBeepEnable;
     uint16_t thermostatSleepTime;
 
-    uint8_t minFanRuntime;
+    uint64_t minFanRuntime;
 
     char *timezone;
     uint16_t timezone_sel;
